+++ packages/fraxel/src/input/index.ts
@@ -1,1 +1,4 @@
+export * from './pointer.js'
+export * from './key.js'
 export * from './input.js'
+export * from './viewport.js'